feature: add Milestone Model
